requirements.txt
aiomysql==0.0.21
hjson==3.0.2
PyMySQL==0.9.3
python-dotenv==0.15.0

shortuuid~=1.0.1